Use `Offset32` as `i32` in ISLE
This commit updates the x64 and aarch64 backends to use the `i32`
primitive type in ISLE when working with an `Offset32` instead of a
`u32`. This matches the intended representation of `Offset32` as a type
which is signed internally and represents how offsets on instructions
are often negative too.

This does not actually change any end results of compilation and instead
is intended to be "just" an internal refactoring with fewer casts and
more consistent handling of offsets.

+++ cranelift/codegen/src/isa/aarch64/lower/isle.rs
@@ -571,16 +571,16 @@ impl Context for IsleContext<'_, '_, MInst, AArch64Backend> {
         }
     }
 
-    fn amode(&mut self, ty: Type, addr: Value, offset: u32) -> AMode {
+    fn amode(&mut self, ty: Type, addr: Value, offset: i32) -> AMode {
         let addr_ty = self.value_type(addr);
         assert!(addr_ty == I64 || addr_ty == R64);
-        lower_address(self.lower_ctx, ty, addr, offset as i32)
+        lower_address(self.lower_ctx, ty, addr, offset)
     }
 
-    fn pair_amode(&mut self, addr: Value, offset: u32) -> PairAMode {
+    fn pair_amode(&mut self, addr: Value, offset: i32) -> PairAMode {
         let addr_ty = self.value_type(addr);
         assert!(addr_ty == I64 || addr_ty == R64);
-        lower_pair_address(self.lower_ctx, addr, offset as i32)
+        lower_pair_address(self.lower_ctx, addr, offset)
     }
 
     fn fp_cond_code(&mut self, cc: &condcodes::FloatCC) -> Cond {
